requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# AI + LangChain stack
langchain==0.3.0
langchain-community==0.3.0
langchain-google-genai==2.0.0
google-generativeai>=0.8.0,<1.0.0

# Data validation
pydantic>=2.4.0,<3.0.0
